Further changes to BE and logic

+++ educouchsystem/src/main/java/com/educouch/educouchsystem/service/AssessmentServiceImpl.java
@@ -2,6 +2,7 @@ package com.educouch.educouchsystem.service;
 
 import com.educouch.educouchsystem.model.Assessment;
 import com.educouch.educouchsystem.repository.AssessmentRepository;
+import com.educouch.educouchsystem.util.exception.AssessmentNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -24,12 +25,22 @@ public class AssessmentServiceImpl implements AssessmentService {
     }
 
     @Override
-    public Assessment retrieveAssessmentById(Long assessmentId) {
-        return assessmentRepository.findById(assessmentId).get();
+    public Assessment retrieveAssessmentById(Long assessmentId) throws AssessmentNotFoundException {
+        Assessment assessment = assessmentRepository.findById(assessmentId).get();
+        if (assessment != null) {
+            return assessment;
+        } else {
+            throw new AssessmentNotFoundException("Assessment Id " + assessmentId + " does not exist!");
+        }
     }
 
     @Override
-    public void deleteAssessment(Long assessmentId) {
-        assessmentRepository.deleteById(assessmentId);
+    public void deleteAssessment(Long assessmentId) throws AssessmentNotFoundException {
+        Assessment assessment = assessmentRepository.findById(assessmentId).get();
+        if (assessment != null) {
+            assessmentRepository.deleteById(assessmentId);
+        } else {
+            throw new AssessmentNotFoundException("Assessment Id " + assessmentId + " does not exist!");
+        }
     }
 }
